requirements.txt
torch~=2.4.1
streamlit~=1.41.1
requests~=2.32.3
pillow~=10.2.0
torchvision~=0.19.1